+++ app/src/main/java/com/shub39/grit/viewmodels/TasksViewModel.kt
@@ -102,6 +102,8 @@ class TasksViewModel(
                         )
                     }
                 }
+
+                is TaskAction.DeleteTask -> repo.deleteTask(action.task)
             }
         }
     }