+++ shared/core/src/commonMain/kotlin/com/shub39/grit/core/habits/presentation/ui/sections/AnalyticsPage.kt
@@ -290,7 +290,7 @@ fun AnalyticsPage(
             onDismissRequest = { editDialog = false },
             onUpsertHabit = { onAction(HabitsAction.UpdateHabit(it)) },
             is24Hr = state.is24Hr,
-            edit = true
+            save = true
         )
     }
 }