+++ shared/core/src/jvmMain/kotlin/com/shub39/grit/core/tasks/presentation/ui/component/TaskUpsertSheet.jvm.kt
@@ -15,9 +15,10 @@ actual fun TaskUpsertSheet(
     categories: List<Category>,
     onDismissRequest: () -> Unit,
     onUpsert: (Task) -> Unit,
+    onDelete: () -> Unit,
     is24Hr: Boolean,
     modifier: Modifier,
-    edit: Boolean
+    save: Boolean
 ) {
     var showDateTimePicker by remember { mutableStateOf(false) }
 
@@ -26,8 +27,9 @@ actual fun TaskUpsertSheet(
         categories = categories,
         onDismissRequest = onDismissRequest,
         onUpsert = onUpsert,
+        onDelete = onDelete,
         is24Hr = is24Hr,
-        edit = edit,
+        save = save,
         notificationPermission = true,
         showDateTimePicker = showDateTimePicker,
         updateDateTimePickerVisibility = { showDateTimePicker = it },