+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 - Made UI adaptive to larger screen sizes
 - Added grit server to manage tasks and habits on desktop in sync with app
 - Made dates clickable on Habits list (#173)
+- Optimised task reordering logic (#182)
+- Added delete button on task edit sheet (#181)
+- Replaced "Edit" with "Save" on task edit sheet (#180)
+- Removed about libraries (#168)
 
 ## 5.1.0
 - Added new all tasks widget